refactor: organize tests by feature (cabins, bookings, shared)
- Reorganize tests into folders: __tests__/cabins, __tests__/bookings, __tests__/shared
- Move CabinCard, CabinFilters, useCabins tests to cabins folder
- Move useBooking tests to bookings folder
- Move shared utilities (test-utils, jest.d.ts, mocks) to shared folder
- Update imports to reflect new structure
- Fix jest.setup.js to properly handle node vs jsdom environments
- Remove React Query global mocks to allow real hook testing

Test Status:
- 36 tests passing
- All test suites passing
- Organized by feature for better maintainability
